Replace hyphen for minus sign only inside math
This commit improves the logic for replacing hyphens by minus signs.
Before, every hyphen was replaced by a minus sign. Now, a hyphen is
replaced by a minus sign only if the hyphen is inside inline math or if
there is no inline math at all.

Examples:

  * `a - b` becomes `a − b` (hyphen replaced)
  * `Multi-dimensional $a - b$` becomes `Multi-dimensional $a − b$`
    (only the second hyphen replaced)

Closes #128

---

As I was working on this issue, I discovered a tiny imperfection inside
`leafmap`, so I also updated that.